Determine if there is a crc computed, and exit prior to recomputing if no overwrite allowed.
putCRC() would compute the new CRC64, and then attempt to write, with a parameter passed to setxattr() to allow/disallow overwriting based on userpolicy.  This resulted in checkit computing the checksum for a file, then warning the user that one already existing and could not be overwritten.

Now, checkit checks for an existing attribute, and if no overwriting is allowed, it stops processing the file.  This makes a significant improvement when storing checksums for large files, where overwriting existing checksums is not allowed and a checksum already exists.